Move rulebook into its own subfolder
